fix(react-theme): Update yellow shared color
