+++ packages/react-theme/src/alias/highContrastColor.ts
@@ -8,11 +8,8 @@ import {
   hcHighlight,
   hcHighlightText,
   hcHyperlink,
-  sharedColors,
-  white,
-  black,
 } from '../global/colors';
-import type { GlobalSharedColors, ColorTokens, ColorPaletteTokens } from '../types';
+import type { ColorTokens } from '../types';
 
 export const generateColorTokens = (): ColorTokens => ({
   colorNeutralForeground1: hcCanvasText, // CanvasText Global.Color.hcCanvasText
@@ -154,22 +151,3 @@ export const generateColorTokens = (): ColorTokens => ({
   colorBrandShadowAmbient: 'rgba(0,0,0,0.30)', // rgba(0,0,0,0.30) undefined
   colorBrandShadowKey: 'rgba(0,0,0,0.25)', // rgba(0,0,0,0.25) undefined
 });
-
-export const colorPaletteTokens: ColorPaletteTokens = (Object.keys(sharedColors) as Array<
-  keyof GlobalSharedColors
->).reduce((acc, sharedColor) => {
-  const color = sharedColor.slice(0, 1).toUpperCase() + sharedColor.slice(1);
-  const sharedColorTokens = {
-    [`colorPalette${color}Background1`]: white,
-    [`colorPalette${color}Background2`]: black,
-    [`colorPalette${color}Background3`]: white,
-    [`colorPalette${color}Foreground1`]: black,
-    [`colorPalette${color}Foreground2`]: white,
-    [`colorPalette${color}Foreground3`]: white,
-    [`colorPalette${color}BorderActive`]: hcHighlight,
-    [`colorPalette${color}Border1`]: white,
-    [`colorPalette${color}Border2`]: white,
-  };
-
-  return { ...acc, ...sharedColorTokens };
-}, {} as ColorPaletteTokens);
