+++ packages/react-theme/src/utils/createTeamsDarkTheme.ts
@@ -1,4 +1,6 @@
-import { colorPaletteTokens, generateColorTokens } from '../alias/teamsDark';
+import { colorPaletteTokens } from '../alias/teamsDarkColorPalette';
+import { generateColorTokens } from '../alias/teamsDarkColor';
+
 import { borderRadius, fontSizes, lineHeights, fontFamilies, strokeWidths, fontWeights } from '../global/index';
 import { createShadowTokens } from './shadows';
 import type { BrandVariants, Theme } from '../types';
